# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AroundU API
  description: |
    AroundU is a local services marketplace connecting clients with skilled workers.
    
    ## Authentication
    Most endpoints require a JWT Bearer token. Obtain one via `POST /api/v1/auth/login`.
    
    ## Rate Limiting
    Certain endpoints enforce rate limits. Exceeding the limit returns `429 Too Many Requests`.
    
    ## Pagination
    List endpoints accept `page` (0-based) and `size` (max 100) query parameters.
    Responses include `totalElements`, `totalPages`, and `last` metadata.
    
    ## Sorting
    Sortable endpoints accept `sortBy`, `sortDirection`, `secondarySortBy`, and `secondarySortDirection`.
    Distance-based sorting is available via `sortByDistance=true` with latitude/longitude coordinates.
  version: 1.0.0
  contact:
    name: AroundU Dev Team
    email: support@aroundu.com
    url: https://aroundu.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:20232
    description: Local development
  - url: https://preprod.aroundu.com
    description: Pre-production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token obtained from POST /api/v1/auth/login
  schemas:
...
    BidCreateRequest:
      type: object
      required: [bidAmount]
      properties:
        bidAmount:
          type: number
          format: double
          minimum: 0.01
          example: 120.00
        partnerName:
          type: string
          example: Helper Co.
        partnerFee:
          type: number
          format: double
          minimum: 0.01
          example: 15.00
        notes:
          type: string
          example: I can start tomorrow morning
    BidHandshakeRequest:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: boolean
          example: true
    BidResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 50
        jobId:
          type: integer
          format: int64
          example: 100
        workerId:
          type: integer
          format: int64
          example: 2
        bidAmount:
          type: number
          format: double
          example: 120.00
        partnerName:
          type: string
        partnerFee:
          type: number
          format: double
        notes:
          type: string
        status:
          type: string
          enum: [PENDING, SELECTED, REJECTED, WITHDRAWN]
          example: PENDING

    # ── Payment ───────────────────────────────────────────────────
    PaymentLockRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          example: 150.00
    PaymentReleaseRequest:
      type: object
      required: [releaseCode]
      properties:
        releaseCode:
          type: string
          example: "REL-XYZ789"
    PaymentResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 200
        jobId:
          type: integer
          format: int64
          example: 100
        clientId:
          type: integer
          format: int64
          example: 1
        workerId:
          type: integer
          format: int64
          example: 2
        amount:
          type: number
          format: double
          example: 150.00
        paymentMode:
          type: string
          enum: [ESCROW, OFFLINE]
          example: ESCROW
        status:
          type: string
          enum: [PENDING_ESCROW, ESCROW_LOCKED, RELEASED, FAILED, OFFLINE]
          example: ESCROW_LOCKED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
paths:
...
  /api/v1/bid/jobs/{jobId}/bids:
    post:
      tags: [Bids]
      summary: Place bid
      description: Worker places a bid on an open job. Rate limited to 20 requests/hour.
      operationId: placeBid
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: workerId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BidCreateRequest'
      responses:
        '201':
          description: Bid placed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidResponseDTO'
        '400':
          description: Validation error or duplicate bid
        '404':
          description: Job not found
        '429':
          description: Rate limit exceeded
    get:
      tags: [Bids]
      summary: List bids for job
      description: Returns all bids submitted for a specific job.
      operationId: listBids
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Bids retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BidResponseDTO'
        '404':
          description: Job not found
  /api/v1/bid/bids/{bidId}/accept:
    post:
      tags: [Bids]
      summary: Accept bid
      description: Client accepts a bid; other bids for the job are rejected.
      operationId: acceptBid
      security:
        - bearerAuth: []
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Bid accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidResponseDTO'
        '403':
          description: Not the job owner
        '404':
          description: Bid not found
        '409':
          description: Bid already processed
  /api/v1/bid/bids/{bidId}/handshake:
    post:
      tags: [Bids]
      summary: Worker handshake
      description: Worker accepts or declines a selected bid.
      operationId: handshakeBid
      security:
        - bearerAuth: []
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: workerId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BidHandshakeRequest'
      responses:
        '200':
          description: Handshake processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidResponseDTO'
        '400':
          description: Invalid handshake state
        '404':
          description: Bid not found

  # ════════════════════════════════════════════════════════════════
  #  PAYMENTS
  # ════════════════════════════════════════════════════════════════
  /api/v1/jobs/{jobId}/payments/lock:
    post:
      tags: [Payments]
      summary: Lock escrow payment
      description: Client locks the escrow amount for a job before work begins.
      operationId: lockEscrow
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentLockRequest'
      responses:
        '201':
          description: Escrow locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseDTO'
        '400':
          description: Validation error
        '404':
          description: Job not found
        '409':
          description: Escrow already locked
  /api/v1/jobs/{jobId}/payments/release:
    post:
      tags: [Payments]
      summary: Release escrow payment
      description: Client releases escrowed funds to the worker after job completion.
      operationId: releaseEscrow
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentReleaseRequest'
      responses:
        '200':
          description: Payment released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseDTO'
        '400':
          description: Invalid release code
        '404':
          description: Job or payment not found
        '409':
          description: Payment not in locked state